+++ sdk/resourcemanager/containerservice/armcontainerservice/models.go
@@ -40,6 +40,13 @@ type AgentPool struct {
 	Type *string
 }
 
+type AgentPoolArtifactStreamingProfile struct {
+	// Artifact streaming speeds up the cold-start of containers on a node through on-demand image loading. To use this feature,
+	// container images must also enable artifact streaming on ACR. If not specified,
+	// the default is false.
+	Enabled *bool
+}
+
 // AgentPoolAvailableVersions - The list of available versions for an agent pool.
 type AgentPoolAvailableVersions struct {
 	// REQUIRED; Properties of agent pool available versions.
@@ -70,6 +77,20 @@ type AgentPoolAvailableVersionsPropertiesAgentPoolVersionsItem struct {
 
 	// The Kubernetes version (major.minor.patch).
 	KubernetesVersion *string
+}
+
+// AgentPoolDeleteMachinesParameter - Specifies a list of machine names from the agent pool to be deleted.
+type AgentPoolDeleteMachinesParameter struct {
+	// REQUIRED; The agent pool machine names.
+	MachineNames []*string
+}
+
+type AgentPoolGPUProfile struct {
+	// The default value is true when the vmSize of the agent pool contains a GPU, false otherwise. GPU Driver Installation can
+	// only be set true when VM has an associated GPU resource. Setting this field to
+	// false prevents automatic GPU driver installation. In that case, in order for the GPU to be usable, the user must perform
+	// GPU driver installation themselves.
+	InstallGPUDriver *bool
 }
 
 // AgentPoolListResult - The response from the List Agent Pools operation.
@@ -150,6 +171,10 @@ type AgentPoolUpgradeSettings struct {
 	// practices, see:
 	// https://docs.microsoft.com/azure/aks/upgrade-cluster#customize-node-surge-upgrade
 	MaxSurge *string
+
+	// The amount of time (in minutes) to wait after draining a node and before reimaging it and moving on to next node. If not
+	// specified, the default is 0 minutes.
+	NodeSoakDurationInMinutes *int32
 }
 
 // AgentPoolWindowsProfile - The Windows agent pool's specific profile.
@@ -305,7 +330,7 @@ type GuardrailsAvailableVersion struct {
 	// REQUIRED; Whether the version is default or not and support info.
 	Properties *GuardrailsAvailableVersionsProperties
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -670,7 +695,7 @@ type ManagedCluster struct {
 	// Resource tags.
 	Tags map[string]*string
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -705,6 +730,14 @@ type ManagedClusterAADProfile struct {
 
 	// The AAD tenant ID to use for authentication. If not specified, will use the tenant of the deployment subscription.
 	TenantID *string
+}
+
+// ManagedClusterAIToolchainOperatorProfile - When enabling the operator, a set of AKS managed CRDs and controllers will be
+// installed in the cluster. The operator automates the deployment of OSS models for inference and/or training purposes. It
+// provides a set of preset models and enables distributed inference against them.
+type ManagedClusterAIToolchainOperatorProfile struct {
+	// Indicates if AI toolchain operator enabled or not.
+	Enabled *bool
 }
 
 // ManagedClusterAPIServerAccessProfile - Access profile for managed cluster API server.
@@ -746,7 +779,7 @@ type ManagedClusterAccessProfile struct {
 	// Resource tags.
 	Tags map[string]*string
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -788,6 +821,9 @@ type ManagedClusterAgentPoolProfile struct {
 	// REQUIRED; Windows agent pool names must be 6 characters or less.
 	Name *string
 
+	// Configuration for using artifact streaming on AKS.
+	ArtifactStreamingProfile *AgentPoolArtifactStreamingProfile
+
 	// The list of Availability zones to use for nodes. This can only be specified if the AgentPoolType property is 'VirtualMachineScaleSets'.
 	AvailabilityZones []*string
 
@@ -829,6 +865,9 @@ type ManagedClusterAgentPoolProfile struct {
 
 	// GPUInstanceProfile to be used to specify GPU MIG instance profile for supported GPU VM SKU.
 	GpuInstanceProfile *GPUInstanceProfile
+
+	// The GPU settings of an agent pool.
+	GpuProfile *AgentPoolGPUProfile
 
 	// This is of the form: /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Compute/hostGroups/{hostGroupName}.
 	// For more information see Azure dedicated hosts
@@ -942,6 +981,12 @@ type ManagedClusterAgentPoolProfile struct {
 	// fail to run correctly. For more details on restricted VM sizes, see:
 	// https://docs.microsoft.com/azure/aks/quotas-skus-regions
 	VMSize *string
+
+	// The status of nodes in a VirtualMachines agent pool.
+	VirtualMachineNodesStatus []*VirtualMachineNodes
+
+	// Specifications on VirtualMachines agent pool.
+	VirtualMachinesProfile *VirtualMachinesProfile
 
 	// If this is not specified, a VNET and subnet will be generated and used. If no podSubnetID is specified, this applies to
 	// nodes and pods, otherwise it applies to just nodes. This is of the form:
@@ -967,6 +1012,9 @@ type ManagedClusterAgentPoolProfile struct {
 
 // ManagedClusterAgentPoolProfileProperties - Properties for the container service agent pool profile.
 type ManagedClusterAgentPoolProfileProperties struct {
+	// Configuration for using artifact streaming on AKS.
+	ArtifactStreamingProfile *AgentPoolArtifactStreamingProfile
+
 	// The list of Availability zones to use for nodes. This can only be specified if the AgentPoolType property is 'VirtualMachineScaleSets'.
 	AvailabilityZones []*string
 
@@ -1008,6 +1056,9 @@ type ManagedClusterAgentPoolProfileProperties struct {
 
 	// GPUInstanceProfile to be used to specify GPU MIG instance profile for supported GPU VM SKU.
 	GpuInstanceProfile *GPUInstanceProfile
+
+	// The GPU settings of an agent pool.
+	GpuProfile *AgentPoolGPUProfile
 
 	// This is of the form: /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Compute/hostGroups/{hostGroupName}.
 	// For more information see Azure dedicated hosts
@@ -1121,6 +1172,12 @@ type ManagedClusterAgentPoolProfileProperties struct {
 	// fail to run correctly. For more details on restricted VM sizes, see:
 	// https://docs.microsoft.com/azure/aks/quotas-skus-regions
 	VMSize *string
+
+	// The status of nodes in a VirtualMachines agent pool.
+	VirtualMachineNodesStatus []*VirtualMachineNodes
+
+	// Specifications on VirtualMachines agent pool.
+	VirtualMachinesProfile *VirtualMachinesProfile
 
 	// If this is not specified, a VNET and subnet will be generated and used. If no podSubnetID is specified, this applies to
 	// nodes and pods, otherwise it applies to just nodes. This is of the form:
@@ -1295,8 +1352,9 @@ type ManagedClusterIngressProfile struct {
 
 // ManagedClusterIngressProfileWebAppRouting - Web App Routing settings for the ingress profile.
 type ManagedClusterIngressProfileWebAppRouting struct {
-	// Resource IDs of the public DNS zones to be associated with the Web App Routing add-on. Used only when Web App Routing is
-	// enabled. All public DNS zones must be in the same resource group.
+	// Resource IDs of the DNS zones to be associated with the Web App Routing add-on. Used only when Web App Routing is enabled.
+	// Public and private DNS zones can be in different resource groups, but all
+	// public DNS zones must be in the same resource group and all private DNS zones must be in the same resource group.
 	DNSZoneResourceIDs []*string
 
 	// Whether to enable Web App Routing.
@@ -1395,6 +1453,11 @@ type ManagedClusterNATGatewayProfile struct {
 
 	// Profile of the managed outbound IP resources of the cluster NAT gateway.
 	ManagedOutboundIPProfile *ManagedClusterManagedOutboundIPProfile
+}
+
+type ManagedClusterNodeProvisioningProfile struct {
+	// Once the mode it set to Auto, it cannot be changed back to Manual.
+	Mode *NodeProvisioningMode
 }
 
 // ManagedClusterNodeResourceGroupProfile - Node resource group lockdown profile for a managed cluster.
@@ -1528,6 +1591,9 @@ type ManagedClusterProperties struct {
 	// The agent pool properties.
 	AgentPoolProfiles []*ManagedClusterAgentPoolProfile
 
+	// AI toolchain operator settings that apply to the whole cluster.
+	AiToolchainOperatorProfile *ManagedClusterAIToolchainOperatorProfile
+
 	// Parameters to be applied to the cluster-autoscaler when enabled
 	AutoScalerProfile *ManagedClusterPropertiesAutoScalerProfile
 
@@ -1593,6 +1659,9 @@ type ManagedClusterProperties struct {
 
 	// The network configuration profile.
 	NetworkProfile *NetworkProfile
+
+	// Node provisioning settings that apply to the whole cluster.
+	NodeProvisioningProfile *ManagedClusterNodeProvisioningProfile
 
 	// The name of the resource group containing agent pool nodes.
 	NodeResourceGroup *string
@@ -1682,11 +1751,6 @@ type ManagedClusterPropertiesAutoScalerProfile struct {
 
 	// Available values are: 'least-waste', 'most-pods', 'priority', 'random'.
 	Expander *Expander
-
-	// Available values are: 'least-waste', 'most-pods', 'priority', 'random'. If multiple expanders are configured, they will
-	// be considered in the order in which they are listed, with the first one being
-	// considered first.
-	Expanders []*Expander
 
 	// If set to true, the resources used by daemonset will be taken into account when making scaling down decisions.
 	IgnoreDaemonsetsUtilization *bool
@@ -1864,7 +1928,7 @@ type ManagedClusterSnapshot struct {
 	// Resource tags.
 	Tags map[string]*string
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -2028,6 +2092,16 @@ type ManagedServiceIdentityUserAssignedIdentitiesValue struct {
 	PrincipalID *string
 }
 
+// ManualScaleProfile - Specifications on number of machines.
+type ManualScaleProfile struct {
+	// Number of nodes.
+	Count *int32
+
+	// The list of allowed vm sizes. AKS will use the first available one when scaling. If a VM size is unavailable (e.g. due
+	// to quota or regional capacity reasons), AKS will use the next size.
+	Sizes []*string
+}
+
 // MeshRevision - Holds information on upgrades and compatibility for given major.minor mesh release.
 type MeshRevision struct {
 	// List of items this revision of service mesh is compatible with, and their associated versions.
@@ -2045,7 +2119,7 @@ type MeshRevisionProfile struct {
 	// Mesh revision profile properties for a mesh
 	Properties *MeshRevisionProfileProperties
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -2077,7 +2151,7 @@ type MeshUpgradeProfile struct {
 	// Mesh upgrade profile properties for a major.minor release.
 	Properties *MeshUpgradeProfileProperties
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -2458,6 +2532,12 @@ type SSHPublicKey struct {
 	KeyData *string
 }
 
+// ScaleProfile - Specifications on how to scale a VirtualMachines agent pool.
+type ScaleProfile struct {
+	// Specifications on how to scale the VirtualMachines agent pool to a fixed size.
+	Manual []*ManualScaleProfile
+}
+
 // Schedule - One and only one of the schedule types should be specified. Choose either 'daily', 'weekly', 'absoluteMonthly'
 // or 'relativeMonthly' for your maintenance schedule.
 type Schedule struct {
@@ -2494,7 +2574,7 @@ type Snapshot struct {
 	// Resource tags.
 	Tags map[string]*string
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -2696,7 +2776,7 @@ type TrustedAccessRoleBinding struct {
 	// REQUIRED; Properties for trusted access role binding
 	Properties *TrustedAccessRoleBindingProperties
 
-	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	// READ-ONLY; Fully qualified resource ID for the resource. E.g. "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}"
 	ID *string
 
 	// READ-ONLY; The name of the resource
@@ -2779,6 +2859,21 @@ type UserAssignedIdentity struct {
 
 	// The resource ID of the user assigned identity.
 	ResourceID *string
+}
+
+// VirtualMachineNodes - Current status on a group of nodes of the same vm size.
+type VirtualMachineNodes struct {
+	// Number of nodes.
+	Count *int32
+
+	// The VM size of the agents used to host this group of nodes.
+	Size *string
+}
+
+// VirtualMachinesProfile - Specifications on VirtualMachines agent pool.
+type VirtualMachinesProfile struct {
+	// Specifications on how to scale a VirtualMachines agent pool.
+	Scale *ScaleProfile
 }
 
 // WeeklySchedule - For schedules like: 'recur every Monday' or 'recur every 3 weeks on Wednesday'.